probe encode: emit taps bf16 (out_dtype) so no fp32 grid → 1-GPU encode
The encode's host-RAM peak was set by the transient fp32 tap grids (biggest
session M2 ~55 GB, all taps ~94 GB) held before save_cache cast to bf16 — over
the 1-GPU host cap once save's cast overshoots. Cast each tap batch to bf16 on
device before CPU accumulation so the fp32 grid never materialises: peak halves
(~47 GB), fits 1 GPU. Bit-identical to the fp32->save_cache path (same rounding,
earlier) — proven by test_encode_subject_tokens_out_dtype_bf16_is_bit_identical.

+++ scripts/neuroprobe/run_pretrain_probe_suite.py
@@ -489,7 +489,7 @@ def run_encode(sessions, tasks, *, ckpt_path, out_dir, cache_untagged_m3=True,
             )
         grids_sf, labels = encode_subject_tokens(
             model, bands, ppe, clip_len_s=PROBE_CLIP_DUR_S, device=device,
-            surfaces=tuple(surfaces), batch_size=batch_size,
+            surfaces=tuple(surfaces), batch_size=batch_size, out_dtype=torch.bfloat16,
         )
         grids = {tap_of[sf]: g for sf, g in grids_sf.items()}
         meta = SessionMeta(
